+++ package/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.7-preview.1] - 2018-11-15
+### Added
+- Fix IK SceneView Gizmo for Unity 2018.3
+- Bumped minimum Unity version for this package to 2018.3
+
 ## [1.0.6-preview.1] - 2018-11-14
 ### Added
 - Fix effector-target name. Effector becomes the transform that moves and the Target becomes the transform to reach.